Improved backend+added authorization

+++ backend/product/views.py
@@ -32,7 +32,7 @@ class ManageUserView(generics.RetrieveUpdateAPIView):
         return self.request.user
 
 @api_view(['GET', ])
-def Product_by_category(request,slug):
+def product_by_category(request,slug):
     product=Product.objects.filter(category=slug,sold=False)
     serializer=ProductSerializer(product,many=True)
 
@@ -40,85 +40,60 @@ def Product_by_category(request,slug):
 
 
 @api_view(['GET', ])
-def Product_all(request):
+def product_all(request):
     product=Product.objects.filter(sold=False)
     serializer=ProductSerializer(product,many=True)
 
     return Response(serializer.data)
 
 
-@api_view(['GET', ])
-def Product_user(request):
-    prefer=Preferences.objects.filter(user=request.user)
-    a=[]
-    
-    for p in prefer:
-        a.append(p.name)
-    print(a)
-    product=Product.objects.filter(category__in=a).filter(sold=False)
-    serializer=ProductSerializer(product,many=True)
 
-    return Response(serializer.data)
-
-
-@api_view(['GET', ])
-def Product_id(request,slug):
+@api_view(['GET','PUT','DELETE' ])
+def product_actions(request,slug):
     product=Product.objects.get(id=slug)
-    serializer=ProductSerializer(product)
+    if (request.user==product.user):
+        if(request.method=='GET'):
+            serializer=ProductSerializer(product)
 
-    return Response(serializer.data)
+            return Response(serializer.data)
 
+        elif(request.method=='PUT'):
+            serializer=ProductSerializer(product,data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
 
-@api_view(['GET', ])
-def delete_product(request,slug):
-    product=Product.objects.get(id=slug)
-    product.delete()
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)  
 
-    return Response({'message': 'Tutorial was deleted successfully!'},status=status.HTTP_204_NO_CONTENT)
+        elif(request.method=='DELETE'):
+            product.delete()
+            return Response({'message': 'Product was deleted successfully!'},status=status.HTTP_204_NO_CONTENT)    
+
+    return Response({'message': 'You are not authorized for this operation!'}, status=status.HTTP_400_BAD_REQUEST)  
 
 
 @api_view(['POST'])
 def add_product(request):
-    serializer=ProductSerializer(data=request.data)
-    print(request.user)
     user=request.user
-    if serializer.is_valid():
-        serializer.save(seller=user,sold=False)
-        return Response(serializer.data,status=status.HTTP_201_CREATED)
 
-    return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+    if(user.is_seller):
+        serializer=ProductSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(seller=user,sold=False)
+            return Response(serializer.data,status=status.HTTP_201_CREATED)
 
+        return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
 
-
-@api_view(['POST', ])
-def add_prefernce(request):
-    serializer=PreferencesSerializer(data=request.data)
-    if serializer.is_valid():
-        serializer.save(user=request.user)
-        return Response(serializer.data,status=status.HTTP_201_CREATED)
-
-    return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
-
-@api_view(['GET', ])
-def User_preference(request):
-    product=Preferences.objects.filter(user=request.user)
-    serializer=PreferencesSerializer(product,many=True)
-
-    return Response(serializer.data)
+    return Response({'message': 'You are not authorized for this operation!'}, status=status.HTTP_400_BAD_REQUEST) 
 
 
-@api_view(['GET', ])
-def delete_preference(request,slug):
-    prefrence=Preferences.objects.get(id=slug)
-    prefrence.delete()
-    return Response({'message': 'Tutorial was deleted successfully!'},status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['GET',],)
-def Purchase(request,slug):
+def purchase(request,slug):
     product=get_object_or_404(Product,id=slug)
     user=request.user
-    purchas=purchase(buyer=user , product=product)
+    purchas=purchase(buyer=user ,product=product)
     purchas.save()
     product.sold=True
     product.save()
